Move bulk of fxrec main back into the direwolf package
